Fixed: documents with unclosed empty tasg weren't parsed correctly.
BeautifulSoup4 uses a different parser, by default, than BeautifulSoup
3.  The new default is an xml parser that adds closing tags in the
wrong places.

Switched to using the html5lib-based parser. This brought its own
pain, as it adds body, html, and head tags if they aren't present. For
some use cases, that might be a good thing, but probably not this one.